pyproject.toml
[project]
name = "assi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
]
